Fix Solaris compiler warnings
This patch does not change any functionality.

+++ apps/snmpping.c
@@ -636,7 +636,7 @@ int main(int argc, char **argv)
         usernameLen = strlen(username);
     }
     if (1 /* !have-testname-arg */) {
-        snprintf(testname, sizeof(testname) - 1, "snmpping-%d", getpid());
+        snprintf(testname, sizeof(testname) - 1, "snmpping-%ld", (long)getpid());
         testname[32] = '\0';
         testnameLen = strlen(testname);
     }
